requirements.txt
appdirs==1.4.4
asgiref==3.5.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fake-useragent==0.1.11
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
lxml==4.8.0
numpy==1.22.3
pandas==1.4.1
parse==1.19.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyee==8.2.2
PyJWT==2.3.0
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.0
tzdata==2021.5
urllib3==1.26.8
w3lib==1.22.0
websockets==10.2
whitenoise==6.0.0
zipp==3.7.0
